+++ st2common/st2common/validators/api/reactor.py
@@ -15,18 +15,26 @@
 
 import six
 
+from oslo_config import cfg
 from apscheduler.triggers.cron import CronTrigger
 
+from st2common import log as logging
 from st2common.exceptions.apivalidation import ValueValidationException
 from st2common.constants.triggers import SYSTEM_TRIGGER_TYPES
 from st2common.constants.triggers import CRON_TIMER_TRIGGER_REF
 from st2common.util import schema as util_schema
 import st2common.operators as criteria_operators
+from st2common.services import triggers
 
 __all__ = [
     'validate_criteria',
-    'validate_trigger_parameters'
+
+    'validate_trigger_parameters',
+    'validate_trigger_payload'
 ]
+
+
+LOG = logging.getLogger(__name__)
 
 allowed_operators = criteria_operators.get_allowed_operators()
 
@@ -51,25 +59,40 @@ def validate_criteria(criteria):
 
 def validate_trigger_parameters(trigger_type_ref, parameters):
     """
-    This function validates parameters for system triggers (e.g. webhook and timers).
-
-    Note: Eventually we should also validate parameters for user defined triggers which correctly
-    specify JSON schema for the parameters.
+    This function validates parameters for system and user-defined triggers.
 
     :param trigger_type_ref: Reference of a trigger type.
     :type trigger_type_ref: ``str``
 
     :param parameters: Trigger parameters.
     :type parameters: ``dict``
+
+    :return: Cleaned parameters on success, None if validation is not performed.
     """
     if not trigger_type_ref:
         return None
 
-    if trigger_type_ref not in SYSTEM_TRIGGER_TYPES:
-        # Not a system trigger, skip validation for now
+    is_system_trigger = trigger_type_ref in SYSTEM_TRIGGER_TYPES
+    if is_system_trigger:
+        # System trigger
+        parameters_schema = SYSTEM_TRIGGER_TYPES[trigger_type_ref]['parameters_schema']
+    else:
+        trigger_type_db = triggers.get_trigger_type_db(trigger_type_ref)
+        if not trigger_type_db:
+            # Trigger doesn't exist in the database
+            return None
+
+        parameters_schema = getattr(trigger_type_db, 'parameters_schema', {})
+        if not parameters_schema:
+            # Parameters schema not defined for the this trigger
+            return None
+
+    # We only validate non-system triggers if config option is set (enabled)
+    if not is_system_trigger and not cfg.CONF.system.validate_trigger_parameters:
+        LOG.debug('Got non-system trigger "%s", but trigger parameter validation for non-system'
+                  'triggers is disabled, skipping validation.' % (trigger_type_ref))
         return None
 
-    parameters_schema = SYSTEM_TRIGGER_TYPES[trigger_type_ref]['parameters_schema']
     cleaned = util_schema.validate(instance=parameters, schema=parameters_schema,
                                    cls=util_schema.CustomValidator, use_default=True,
                                    allow_default_none=True)
@@ -81,5 +104,48 @@ def validate_trigger_parameters(trigger_type_ref, parameters):
         # allows arbitrary strings, but not any arbitrary string is a valid CronTrigger argument
         # Note: Constructor throws ValueError on invalid parameters
         CronTrigger(**parameters)
+
+    return cleaned
+
+
+def validate_trigger_payload(trigger_type_ref, payload):
+    """
+    This function validates trigger payload parameters for system and user-defined triggers.
+
+    :param trigger_type_ref: Reference of a trigger type.
+    :type trigger_type_ref: ``str``
+
+    :param payload: Trigger payload.
+    :type payload: ``dict``
+
+    :return: Cleaned payload on success, None if validation is not performed.
+    """
+    if not trigger_type_ref:
+        return None
+
+    is_system_trigger = trigger_type_ref in SYSTEM_TRIGGER_TYPES
+    if is_system_trigger:
+        # System trigger
+        payload_schema = SYSTEM_TRIGGER_TYPES[trigger_type_ref]['payload_schema']
+    else:
+        trigger_type_db = triggers.get_trigger_type_db(trigger_type_ref)
+        if not trigger_type_db:
+            # Trigger doesn't exist in the database
+            return None
+
+        payload_schema = getattr(trigger_type_db, 'payload_schema', {})
+        if not payload_schema:
+            # Payload schema not defined for the this trigger
+            return None
+
+    # We only validate non-system triggers if config option is set (enabled)
+    if not is_system_trigger and not cfg.CONF.system.validate_trigger_payload:
+        LOG.debug('Got non-system trigger "%s", but trigger payload validation for non-system'
+                  'triggers is disabled, skipping validation.' % (trigger_type_ref))
+        return None
+
+    cleaned = util_schema.validate(instance=payload, schema=payload_schema,
+                                   cls=util_schema.CustomValidator, use_default=True,
+                                   allow_default_none=True)
 
     return cleaned